# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Cursos de Finanzas",
        "description": "Documentación de la API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/modulos": {
            "get": {
                "tags": [
                    "Módulos"
                ],
                "summary": "Listar módulos con filtros opcionales",
                "operationId": "38bcd20b7ccaefe76569335ef83daa3c",
                "parameters": [
                    {
                        "name": "id_curso",
                        "in": "query",
                        "description": "Filtrar por ID del curso",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "nombre",
                        "in": "query",
                        "description": "Filtrar por nombre del módulo",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "activo",
                        "in": "query",
                        "description": "Filtrar por estado activo/inactivo",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad por página",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de módulos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Módulos"
                ],
                "summary": "Crear un nuevo módulo",
                "operationId": "5128f78082469d924390db2287338fb9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_curso",
                                    "nombre",
                                    "fecha_fin"
                                ],
                                "properties": {
                                    "id_curso": {
                                        "type": "integer"
                                    },
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "fecha_fin": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Módulo creado exitosamente"
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "401": {
                        "description": "Usuario no autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/modulos/{id}": {
            "get": {
                "tags": [
                    "Módulos"
                ],
                "summary": "Obtener un módulo por ID",
                "operationId": "b1fbd2fa61885631213ce5123f25c997",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del módulo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Módulo encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Módulos"
                ],
                "summary": "Actualizar un módulo existente",
                "operationId": "fa3949e1085302c3d7dca6fd7fd67122",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del módulo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_curso": {
                                        "type": "integer"
                                    },
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "fecha_inicio": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "fecha_fin": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "activo": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Módulo actualizado exitosamente"
                    },
                    "404": {
                        "description": "Módulo no encontrado"
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "401": {
                        "description": "Usuario no autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Módulos"
                ],
                "summary": "Eliminar lógicamente un módulo",
                "operationId": "dbc4449e3f0d2a55bf4597e0f44b0a58",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del módulo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Módulo eliminado exitosamente"
                    },
                    "409": {
                        "description": "Módulo no se puede eliminar por estar asignado a inscripciones"
                    },
                    "404": {
                        "description": "Módulo no encontrado"
                    },
                    "401": {
                        "description": "Usuario no autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Introduce el token JWT aquí. Ejemplo: Bearer {token}",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}